Quite an update, I'd say. For one thing, Google Maps is not working rigth now but I'll get to that later. However, on the flipside, I managed to create my very first service! (EEEK!)!!
The service is extremely simple, if it receives an intent to
START_NOTIFICATION, it will do so. Not just that, but I also implemented
ServiceTools, which is an awesome new utility which will start any
service (reflectively) and use a callback to properly fill out the
intent. Of course I can't take all of the credit myself, gotta thank the
person from Stack Overflow.

Lastly, created my own OnBoot BroadcastReceiver, my very first. All it
does it start the service, which causes the service to start the
notification. Yippee!