Broke quite a few things trying to improve this, spent literally all day on it, still not finished, but getting closer...
1) Changed the TitleBar and am in the process of adding a BottomBar,
where the title bar can be dragged from the text and empty portion, and
includes an options menu and a close menu. The bottom bar is meant for
resizing, although I'm not having much success (or fun, literally spent
last 6 hours on it, dammit!) it's coming along like a tortoise in a
marathon. Painfully slow.

2) Dragging/Moving any widget to the right side of the screen will cause
it to minimize, icon and all. I'll also implement the inverse, dragging
the icon to a portion of the screen to spawn it there. Hopefully
tomorrow.