Very cool update. But first, I must say that today in class, during presentations, I was beginning to think there wouldn't be any real "competition" for me (which truth be told, so far there wasn't), until one person blew everyone away. It started out dry, but as he got into the specifics, it was actually extremely impressive, and he did it all in one week. That takes extreme dedication.
Anyway, back to my project. I started implementing screen capture, and
it's actually going pretty well, I already am planning out the layout,
and already have a way to capture the screen, and I am going to be
adding a floating icon and/or notification to allow users to end the
recording with a press of a button. Simply amazing, I say.