Had a drunken epiphany as to how to improve upon everything, and boy oh boy does it look good.
1) There will be a top and bottom bar while than a horizontal one.
Hence, making it more friendly for portrait orientation and making it
more user-friendly for desktop/laptop users.
2) The overall philosophy has somewhat changed. For now, the target is
to provide a multi-tasking window-manager to Android as best as
possible, and hopefully make transitioning in and out of the app as
seamless as possible. Not only is it an all-in-one hub, it is also a
home-away-from-home and in a cozy, comfortable, and familiar
environment.
3) Started planning on MenuBuilder, which will help ease along the
process of creating new menus by filling it out for myself ahead of time
in a generic and loosely coupled way.
4) Fixed that annoying issue where ScreenRecorderFragment did not update
the state the second time around it was created.

Planning on making extra changes...

1) Change FloatingFragment to WindowFragment or DynamicWindowFragment
2) Make more use of RxJava (Dunno how tho :( ~)