I did a lot of work on improving the efficiency of everything. Unfortunately I couldn't use RxView for the main touch event handling for FloatingFragments as it is extremely paranoid as to calls on non UI thread, which makes sense, but it just doesn't fit my solution.
Anyway, I correctly managed to defer work to a background thread for
calculating the new sizes and coordinates of the view, but unfortunately
resize isn't working as intended so I probably have to spend another day
on that.

I DO however see the performance benefits already. What used to be
sluggish is actually now very smooth. FloatingFragments lag behind of
course, however. However, I can see that it will be possible using
RxJava to debounce and handle the backpressure of excess touch events.
Yes, I can see this working wonderfull.