Major documentation update and of course a ton of bug fixing.
1) Maximization, Snapping, etc, work again, but then again they worked
last update too. Now they work AND look better
1A) Updated MeasureTools and FloatingFragment to be a lot more
documented and readable for the professor when he grades my project. I
don't regret it at all, because IMO it looks a LOT better now.
2) Added the Info bar, which lets the user know of the current progress
of the program. I.E, on start if the program is busy inflating and
deserializing each view, it lets the user know it is "Restoring
Session", otherwise it is "Creating Session".
2A) Developed a protoype for the Mac OS X style menu options, as now the
current FloatingFragment selected shows up in the Info bar. Imagine that
that's where each menu's specific option will go.
3) Optimized a few things here and there.