Began implementing my new idea for better serialization. For one thing, I will no longer have to worry about session data being corrupted anymore because everything will be written to a local SQLite database. Normally I'd think this is overkill, but in fact it is rather necessary.
The old model, using JSON, unfortunately while it did allow me to
create, delete and read in session data, it was impossible to update
without doing so all at once, which is easy to see why it can fail, and
relying on an AsyncTask was also another issue.

My plan right now is have a SQLite database which gets updated whenever
a view's state changes, however of course this would be combersome on
the UI thread, I plan on having a Session Manager, a handler thread
which will handle any and all updating of the database.

Hence, even if this is being called very often, and even if changes are
made almost incomplete (as GSON is used to serialize, it may be possible
that GSON serializes outdated data), at least it will remain consistent.
Overall, I think it's going along just fine. Very tired right now from
Finals week however, so I'm more preparing the abstract. Anyway, since
it's local there's no issue about sending information too often as it's
all on disk, and SQLite is optimized for things like this.