Managed to finish up the SessionManager and increase efficiency of moving and resizing views. For one, I no longer check at all until user lets go, however I do rebound for the resize operation as setting the width and/or height will cause the view to go to extreme numbers, like the maximum integer number possible. Which is pretty insane, but ah well.
I also refactores FloatingFragment to do all processing, and renamed
ViewController to ViewState so it can be more easily serialized. Also,
mainly because I want to mimic the MVC-pattern as much as possible. As
far as I can tell, the FloatingFragment acts as the Controller and the
ViewState acts as the Model. Not 100% certain however.