Okay, finally, FINALLY managed to get the whole MenuBuilder working. It actually works now, in that it will display all options (although seperator isn't completed) and even their onClick callback can be executed.
I will be making a few changes however... for one,
View.onClickListener() isn't really necessary and a simple interface or
a Runnable can accomplish it just as easily. Then, there's the seperator
and the MenuOptionsAdapter as a whole. It needs an overhaul, because it
looks REALLY ugly right now, although the syntax looks great. Finally, I
need some way to get PopupWindow to be an appropriate size, as setting
it statically is a VERY bad idea, and it needs to be scaled down.