Began creating the Resizable and Moveable dynamic components the user may create for each. As always it's a rocky start, but with the gradual changes and progress, it can become very big. I can already see how it will look awesome when finished. I'll guideline the overall philosophy I'm attempting to achieve here...
1) Be able to move and resize components dynamically
2) Save layouts as either complete or as drafts
3) Reflectively add different types of view children to a dynamic
component which handles and abstracts the resizing and movement. These
will be recreated later through serialization/deserialization as XML or
JSON.

For now, this is a milestone goal and does not reflect the end goal for
everything, however WHEN it makes it to that stage, it'll be even
presentable and complete the proof-of-concept and I can go work on the
buffering portion, and I can go back and forth between each until they
meet the stable requirements to be merged with Master.