Began implementing the serialization and deserialization of the DynamicComponents. I created an interface which does the following...
A) Creates the view, as it needs to be inflated ONLY when build() is
called, and it needs to be added to the ViewGroup inflated later as
well. Any setting up of callbacks and whatnot need to be initialized
there too.
B) Obtains the result from the view. If it is null, it will not both
checking if it is valid.
C) With that result, if not null, check if it is valid.
D) If it is not valid, return a String embodying the error message.
E) Handle the result if it is valid.

The above interface will alow the builder more dynamic and I suppose
complex building of a menu.

I also am currently decided on JSON rather than XML as I am more
familiar with it, plus I can't see standard XML working for the more
complicated/scripted stuff, like buttons, clicks, etc.