Deleted some older stuff, and began thinking/planning out the implementation of the buttons. I originally was going to use Javax.script.* but apparently Android does not implement that, hence that's out of the option. What I can say, however, is that I found a different alternative. The ImplementationsIdeas.md should detail my thought process, however to summarize, what I will be doing is that I need to create a helper menu which will find any and all references/components/variables in the current scope, obtain their possible actions, conditionals, etc. It needs to be robust and smart enough to identify and allow the user to easily do whatever they could possibly want.
Not only does this mean I get to create a Domain-Specfic Language, which
I would think would be rather invaluable for me as a potential software
engineer. I also would get to experiment even more with UI/UX, as well
as low-level, making me an all-around well-balanced full-stack Android
developer :P

Also, I plan on overhauling the component menu as well as the widget
selection menu. I'm beginning to think of a simple heirarchy for it as
well.

View
|
--> Text
    |
    --> Button
    |
    --> Editable
    |
    --> Immutable
|
--> Image
    |
    --> Button
    |
    --> View
|
--> Web
    |
    --> View
|
--> Grouping
    |
    --> Layout
    |
    --> RadioButtons
    |
    --> CheckBoxes