Implemented the DrawerLayout as well as a simple builder and adapter utilizing an ExpandableListView. Was actually rather easy with a nice example to copy/modify from.
I also moved the creation/selection of components to the left hand side,
and the attributes to the right hand side. They look relatively nice
right now, MUCH better than a modal dialog. Even created a subclass of
DrawerLayout so that it allows me to move and resize while the attribute
drawer is open.

Doesn't mean I'm anywhere near done, didn't start the button/DSL yet,
but I'm paving the way for it since the modal dialog can solely be used
to create the script.