A few things...
1) Refactored everything so that SAK-Overlay is split into three parts,
a CORE, a WIDGETS, and a COMPONENT. Core encapsulates things which both
WIDGETS and COMPONENT use as well as the core essentials to begin the
project (I.E, MainActivity and NotificationService, and the database).
WIDGETS contains all Widget related classes and data, while Component
handles all Component classes and data. This is going to be a LOT easier
in the long run because now I can easily tell which belongs to which and
have everything be more organized. It's not finished, in terms of
organization, but it's an improvement.
2) Began working on how I'm going to be implementing the ReferenceTypes.
Each ReferenceType contains it's ConditionalType (holding Conditional
methods), and ActionType (holding Action methods). Paired with
reflection and generics, this should make it a lot eaiser and do-able
when it comes to obtaining a list of ALL references, and receiving the
conditionals/actions of each particular reference. HOw I'm going to be
doing that, I'm still deciding. Still good progress.