Began finalizing a bit of the ReferenceType, ActionType and ConditionalType. It is beginning to actualy look very good/elegant in design. To give an example, let's say that we wish to obtained a list of all references names, needed to show to the user. Right now this part isn't implemented but it would be trivial to do so. It would look like such...
Stream.of(helper.getAllReferences())
	.map(ref -> ref.getName())
	.forEach(adapter::add);

Look at how simple and elegant it looks. Now, imagine a more complex
query, such as "List all conditional methods by method name for each
reference." It would also be simple...

Stream.of(helper.getAllReferences())
	.flatMap(ref -> Stream.of(ref.getConditionals().getAllMethods())
	.map(MethodWrapper::getMethodName)
	.forEach(adapter::add);

Extremely simple and elegant, despite the name of the methods, which can
be made shorter. Regardless, doing it this way has made it so that it's
abstract and elegant enough that complex queries can be done in a few
lines of code with the help of Lambdas, Streams and Observables.

Also, I added some static methods which can abstract away the work of
creating Conditionals and Actions, and the Reference itself. For
example...

ReferenceType<BaseComponent> ref = ReferenceType.from(BaseComponent.this,
BaseConditionals.class, BaseActions.class);

Simple.