I did it! Wunderbar! Everything is actually working now. Now, not to say that everything works, per say, but it is presentable enough to get across what I am attempting to accomplish. I am however impressed with how I am managing this however. It's going like a dream. It's a very complex and sophisticated feature(s) I'm trying to implement and I'm designing them elegantly and they are reusable. Oh how I love programming.
Since this will be committed back to Master, as I haven't done that in a
while, I will list below the changes made.

1) Separated significant parts of this project into modules, Core,
Components, and Widgets. Doing so has greatly increased cleanliness of
the code and makes it easier to track down specific bugs.
1A) Core manages the classes and resources needed by all modules, I.E
the main database, the main activity, etc.
1B) Widget manages the Widget sections, like Web Browser, Google Maps,
and the Widget specific database module.
1C) Component is for layout creation and affiliated components and
classes.
2) Created and began implementing a way to create custom layouts. It is
a drag and drop tool thus far, but also allows user to modify attributes
with AttributeEditor/Menu. There is also a simple implementation for
creating callbacks through predefined combination of sorts. Right now
it's underdeveloped, but I am working on it daily, so it's getting
there.