In the past two days, I've been thinking of more improvements, the first being that the user will want to set the background image from some file, or online resource. Of course, to do so, means that I need to find some way to allow the user to select a file in-app, as launching other activities isn't satisfactory, and because the user may want to do some File IO themselves, hence I decided to create my own minimal filesystem browser, starting at the root of the sdcard folder.
Right now it's still in early planning stages, however, it is coming
along relatively well in how it is going to be done. I am making strong
use of Observables this time around, making asynchronous processing and
retrieval of files (and wrapping said files), and even sorting them
(Yay, made a new Transformer!!!) and handling the results on the UI
thread in a most elegant manner.

Secondly, I plan on overhauling LineOfCode to allow optional operators
and even modifiers. Operators will be like such: { "AND", "OR" }, a
Modifier is like such: { "NOT" }. Of course there will be more. For
example, TypedOperators, will display operators depending on the
preceeding ReferenceType's type, once Getters and Setters are
initialized. For example, I internalized it like this...

if EditText0.getInput() < 100 AND EditText0.getInput() > 0

The above is a combination of Reference.Conditional() pair,
TypedOperator (as the getInput() returns a ReferenceType based on
EditText input type, which encapsulates something similar to dynamic
typing), and the user input, the user input can be delegated for later
(will get to that later), or handled now in a AlertDialog window. It can
get complicated real quick, but abstractions help that a lot.

I should also be able vastly improve on both by the end of the month. In
fact, I expect to see vast steady improvements.