Okay, so finally back into working on this. I decided to take a break, as I realize it's the exhaustion of working on the same thing, and besides, there is a TON of things which needs to be redone. Not just adding streams and further improving code quality, but overall design too.
I've begun creating a separate AbstractWidget, which will be the new
BaseWidget once it is finished. The reasoning is that the BaseWidget
does WAY too much. Instead I've begun implementing a WidgetView that
will handle it's own movement, resizing, and snapping. The
AbstractWidget is also no longer a Fragment (as there was no need to
begin with, only for a grade).

I've also changed my documentation style, using the beautiful looking
structured comments to describe and separate code into sections. MUCH
needed, as well.